Classes HomeController and UserController moved to com.negotium.negotiumapp.controller
Converted HomeController into Rest,
Converted UserController into Rest,

Reformatted code of DuplicateEmailException class,
Reformatted code of DuplicateUsernameException class,
Reformatted code of InvalidLoginException class,
Reformatted code of User class,
Reformatted code of UserRepository class,
Reformatted code of UserRoleRepository class,
Rearranged autowiring type in UserService class,
