+++ src/main/java/com/negotium/negotiumapp/exception/DuplicateEmailException.java
