+++ src/main/java/com/negotium/negotiumapp/exception/DuplicateUsernameException.java
